Convert KeyboardAndMouseStatistics for DI
Refactor renamed KeyboardAndMouseStatistics to
KeyboardAndMouseStatisticsImpl and added an abstracted interface in
order to split interface from implementation. A factory then allows us
to slik inject within the consumers (GraphStatsController and
TextStatsController) and create instances of the abstracted class.
Later, we can more easily unit test the controllers by injecting a
mocked factory that will create mocks of the KeyboardAndMouseStatistics.

Issue: #27