Create ExpiringCircularBuffer data structure
The ExpiringCircularBuffer is an insert and expiration-optimized data
structure meaning, inserts are O(1) and expirations are O(1). This
is optimal for capping potentially large influxes of data, but also
expire the data if it's stale. This data structure will be used to store
keyboard/mouse input times to calculate APM among other metrics.