Create AccumulatingCircularBuffer
This data structure accumulates added values over a given timespan into
a single element, shifting the element after the given timespam has
elapsed. In database terms, this is an aggregation, grouped by time.

The use of this data structure is to keep track of true APM (recorded
actions over the span of a minute) for a certain amount of time. The
accumulation time in this case is 60 seconds, and the size of the buffer
will be determined later.